Module Asllib

module AST : sig ... end

module ASTUtils : sig ... end

This module provides some tools to work on ASL ASTs.

module Backend : sig ... end

module Bitvector : sig ... end

This module provide an interface to ASL bitvector, and main operations on it.

module Builder : sig ... end

Builds an AST.t from some files.

module Diet : sig ... end

An interval set implementation using diets.

module Env : sig ... end

The runtime environment used by Interpreter.

module Error : sig ... end

Error handling for Asllib.

module Gparser0 : sig ... end

module Instrumentation : sig ... end

Provide some instrumentation backends for Interpreter and Typing.

module Interpreter : sig ... end

module Lexer : sig ... end

module Lexer0 : sig ... end

ASL lexer support

module Native : sig ... end

module PP : sig ... end

Pretty-printers for ASL ASTs.

module Parser : sig ... end

module Parser0 : sig ... end

module RepeatableLexer : sig ... end

module Serialize : sig ... end

This module converts an AST into a valid ocaml string that represents it.

module SimpleLexer0 : sig ... end

module Splitasl : sig ... end

Split (ASL) lexbuffer at "// =======..." limits

module StaticEnv : sig ... end

Static Environment used for type-checking (cf Typing.

module StaticInterpreter : sig ... end

module Storage : sig ... end

This module produces an map like interface to model memory storage.

module Types : sig ... end

Type Algebra

module Typing : sig ... end

The Typing module is yet a single-entry-point module. It only exports the function annotate_ast which fills type-annotations holes in the AST. It should provide enough information to disambiguate any type-dependent behaviour.

module Version : sig ... end