Module Interpreter.Make

Parameters

module B : Backend.S

module C : Config

Signature

module B = B

val run_env : (AST.identifier * B.value) list -> B.ast -> B.value B.m

run env0 ast runs the function main of the ast, in an environment build from the ast. However, the (global) identifiers listed in the A-list env0 will take their initial values from env0 and _not_ from ast.

val run : B.ast -> B.value B.m

Shorthand for run [] ast

val run_typed : B.ast -> StaticEnv.env -> B.value B.m

run_typed ast env runs the function main of the typed-checked ast, in typed-checking environment env.