Module Types.Domain

The domain of a type is the set of values which storagbe element of that type may hold.

type t

Abstract value set.

val pp : Stdlib.Format.formatter -> t -> unit

A printer for the domain type.

val of_type : env -> AST.ty -> t

Construct the domain of a type.

val mem : AST.literal -> t -> bool

mem v d is true if and only if v is in d.

val equal : t -> t -> bool

Wheather two domains are equal.

val compare : t -> t -> int option

The inclusion order on domains.

It is a partial order.